+++ src/main/java/com/ourkitchen/data/repository/KitchenInfoRepository.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.ourkitchen.data.entity.KitchenInfoEntity;
 
 @Repository
-public interface KitchenInfoRepository extends JpaRepository<KitchenInfoEntity, Long>{
+public interface KitchenInfoRepository extends JpaRepository<KitchenInfoEntity, Integer>{
 	public List<KitchenInfoEntity> findByNameContaining(String name);
+	
 }
